added LU decomposition
